+++ strongMan/apps/connections/urls.py
@@ -4,7 +4,8 @@ from . import views
 
 app_name = 'connections'
 urlpatterns = [
-    url(r'^$', views.ChooseTypView.as_view(), name='connections_choose'),
+    url(r'^$', views.overview, name='index'),
+    url(r'^add/$', views.ChooseTypView.as_view(), name='connections_choose'),
     url(r'^create/$', views.create, name='connection_create'),
     url(r'^(?P<id>\d+)/$', views.update, name='connection_update'),
     url(r'delete/(?P<id>\d+)/$', views.delete_connection, name='connection_delete'),
