+++ strongMan/templates/base.html
@@ -25,7 +25,7 @@
                 <span class="icon-bar"></span>
                 <span class="icon-bar"></span>
             </button>
-            <a class="navbar-brand" href="{% url "index" %}">
+            <a class="navbar-brand" href="{% url 'index' %}">
                 <img alt="StrongMan" src="{% static 'img/swanicon.png' %}"> StrongMan
             </a>
         </div>
